Refactor: split app.py into store, hotkeys, and tray modules - Extract data persists and SSE (store.py), hotkey management (hotkeys.py), and webview/tray logic (tray.py) from app.py to improve maintainability. - app.py now contains only Flask routes and startup logic
